Fix shim implementation of `hash_pbkdf2` to work properly when `length` is `0`
According to the PHP docs, when length is set to `0` (as it should be by default when not specified), the length returned should be the size of the output from
the specified hash algorithm.

This fixes the implementation to allow this, and makes the last two arguments optional as they are supposed to be.

Fixes #1